requirements.txt
tensorflow
numpy
opencv-python
scikit-learn
matplotlib
seaborn
fastapi
uvicorn
pillow